Clean up code and tests
